created a settings menu for fallback mode/region change also houses option to enable svchax
