+++ source/menu_settings.h
@@ -1,0 +1,14 @@
+#pragma once
+#include <fstream>
+#include <3ds.h>
+
+#include "svchax/svchax.h"
+#include "menu.h"
+#include "common.h"
+#include "display.h"
+#include "utils.h"
+
+void menu_settings();
+void action_toggle_install();
+void action_toggle_region();
+void action_enable_svchax();
